Changed functionality to float data

+++ app/src/main/java/com/chisw/switchmymind/network/model/InputData.java
@@ -12,37 +12,37 @@ import java.util.List;
 public class InputData {
     @SerializedName("data_1")
     @Expose
-    private List<Byte> data_1;
+    private List<Float> data_1;
 
     @SerializedName("data_2")
     @Expose
-    private List<Byte> data_2;
+    private List<Float> data_2;
 
     @SerializedName("data_3")
     @Expose
-    private List<Byte>data_3;
+    private List<Float>data_3;
 
-    public List<Byte> getData_1() {
+    public List<Float> getData_1() {
         return data_1;
     }
 
-    public void setData_1(List<Byte> data_1) {
+    public void setData_1(List<Float> data_1) {
         this.data_1 = data_1;
     }
 
-    public List<Byte> getData_2() {
+    public List<Float> getData_2() {
         return data_2;
     }
 
-    public void setData_2(List<Byte> data_2) {
+    public void setData_2(List<Float> data_2) {
         this.data_2 = data_2;
     }
 
-    public List<Byte> getData_3() {
+    public List<Float> getData_3() {
         return data_3;
     }
 
-    public void setData_3(List<Byte> data_3) {
+    public void setData_3(List<Float> data_3) {
         this.data_3 = data_3;
     }
